# docs/Mark-Layer2-API-Contracts.yaml
openapi: 3.0.3
info:
  title: MARK Layer-2 and Wellbeing API Contracts
  version: 1.0.0
  description: >-
    Contract-first API specification for MARK workplace assistant tools and
    wellbeing reminder operations.
servers:
  - url: /api/v1
    description: Base path
...
paths:
...
  /integrations/calendar/availability:
    post:
      tags: [Integrations]
      summary: Find available slots for attendees
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarAvailabilityRequest'
      responses:
        '200':
          description: Matching slots and conflict hints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarAvailabilityResponse'
  /integrations/calendar/events:
    post:
      tags: [Integrations]
      summary: Create calendar event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarEventRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCalendarEventResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CalendarAvailabilityRequest:
      type: object
      required: [attendees, from, to, durationMinutes]
      properties:
        attendees:
          type: array
          minItems: 1
          items:
            type: string
            format: email
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        timezone:
          type: string
          default: UTC
        durationMinutes:
          type: integer
          minimum: 15
          maximum: 240
    AvailabilitySlot:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        score:
          type: number
          format: float
        reason:
          type: string
    CalendarAvailabilityResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilitySlot'
        conflicts:
          type: array
          items:
            type: string
    CreateCalendarEventRequest:
      type: object
      required: [title, start, end, attendees]
      properties:
        title:
          type: string
        description:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        timezone:
          type: string
          default: UTC
        attendees:
          type: array
          items:
            type: string
            format: email
        location:
          type: string
        idempotencyKey:
          type: string
    CreateCalendarEventResponse:
      type: object
      properties:
        eventId:
          type: string
        provider:
          type: string
        joinUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [scheduled, pending]
...